backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Validation / Settings ──────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0

# ── Security ───────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ── AI Providers ───────────────────────────────────────────────
anthropic==0.40.0
openai==1.54.0

# ── Document Processing ────────────────────────────────────────
python-docx==1.1.2

# ── Utils ──────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==9.0.0
